ajout de la possibilité de supprimer des listes en mode  read/writer

+++ src/app/home/home.page.ts
@@ -18,6 +18,7 @@ import { Plugins } from "@capacitor/core";
 import { Todo } from "../models/todo";
 import { THIS_EXPR } from "@angular/compiler/src/output/output_ast";
 import { element } from "protractor";
+import { mainModule } from "process";
 const { Storage } = Plugins;
 const { SpeechRecognition } = Plugins;
 const { TextToSpeech } = Plugins;
@@ -135,7 +136,16 @@ export class HomePage implements OnInit, OnDestroy {
   }
 
   public deleteList(list: List) {
+    if (list.owner == this.currentUser.email){
     this.listService.Delete(list);
+    }else if (list.canWrite.indexOf(this.currentUser.email) > -1) {
+ list.canWrite.splice(list.canWrite.indexOf(this.currentUser.email),1)
+      this.afs.collection("lists").doc(list.id).update({ canWrite: list.canWrite });
+    }else if (list.canRead.indexOf(this.currentUser.email) > -1 ){
+      list.canRead.splice(list.canRead.indexOf(this.currentUser.email));
+      this.afs.collection("lists").doc(list.id).update({ canRead:  list.canRead});
+    }
+    this.lists.splice(this.lists.indexOf(list),1);
   }
 
   public toggleColourTheme() {
